requirements.txt
viktor==14.22.0
python-dotenv
openai
numpy>=2.3.1
openseespy==3.3.0.1
pydantic>=2.11.7
kaleido==0.2.1; platform_system == "Linux"
kaleido==0.1.0.post1; platform_system == "Windows"
instructor
plotly